fix: skip unknown assets AB#18920

+++ services/API-service/src/api/point-data/point-data.service.ts
@@ -246,6 +246,9 @@ export class PointDataService {
           countryCodeISO3: assetFids.countryCodeISO3,
         },
       });
+      if (!asset) {
+        continue;
+      }
 
       // Delete existing entries with same date, leadTime and countryCodeISO3 and stationCode
       await this.pointDataDynamicStatusRepo.delete({
